backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
ultralytics==8.1.2
python-dotenv==1.0.0